Move functions used only by the SSPI auth method to sspi_windows.go
